Fixed Movement
It now properly checks if the ball is grounded, although it doesn't work on slopes.

Moving left or right will accelerate the ball until a top speed is reached.

The ball will also slow down fast when the player isn't pressing A or D and the ball is grounded.

I also put each static object in the "Ground" layer. This is what's used by the ball to know what counts as the ground.